requirements.txt
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
Pillow==9.5.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytz==2023.3
sqlparse==0.4.4
requests
firebase-admin